+++ src/pages/post/Post.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <link rel="stylesheet" type="text/css" href="../assets/style/post.css">
+    <link rel="stylesheet" type="text/css" href="../../assets/style/post.css">
     <title>Community clone-emily</title>
 </head>
 <body>
@@ -75,9 +75,9 @@
         </div>
     </main>
     
-    <script type="module" src="../components/header.js"></script>
-    <script type="module" src="../components/modal.js"></script>
-    <script type="module" src="../components/commentsView.js"></script>
-    <script type="module" src = "../scripts/post.js"></script>
+    <script type="module" src="../../components/header.js"></script>
+    <script type="module" src="../../components/modal.js"></script>
+    <script type="module" src="./commentsView.js"></script>
+    <script type="module" src = "./post.js"></script>
 </body>
 </html>
